Move ROS interface to separate module
